refactored form into smaller componenents. Broke up tests accordingly
